--- src/client_conn.h
@@ -1,11 +1,0 @@
-void tetris_send_message (char * message);
-
-void tetris_connect(char * host, int port);
-
-void tetris_disconnect();
-
-void tetris_translate(int x, int y);
-
-void tetris_rotate(int theta);
-
-void tetris_drop();
