--- src/test_player.c
@@ -1,8 +1,0 @@
-#include "player.h"
-
-int
-main(void)
-{
-  player_init();
-  player_create(0, "George");
-}
